requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.5
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
mysqlclient==2.1.1
packaging==21.3
platformdirs==2.5.2
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.16.5
Werkzeug==2.2.2
WTForms==3.0.1
